% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scheme.R
\docType{data}
\name{scheme}
\alias{nil}
\alias{scheme}
\title{dottedpaiR
Scheme functions in R}
\format{\preformatted{ NULL
}}
\usage{
nil
}
\arguments{
\item{ls}{list}

\item{f}{function}

\item{l}{list}

\item{s}{list}

\item{x}{number}

\item{a}{number}

\item{d}{number}
}
\description{
dottedpaiR
Scheme functions in R
}
\examples{
cons(3,cons(4,nil))
cdr(cons(3,cons(4,nil)))
}
\keyword{datasets}

